Beginnings of tests for subscription payments.
These tests verify a TransactionSearch for ProfileID, and also serve as
an example of the method to find subscriptions using that method.

I've also added to the test a creation of an HTML form that can be used
to create a transaction that will allow the test to pass.  This is a
hack, not unlike the hacks existing in the tests now that ask for
pasting of a transaction ID.  Long term, if someone wants to do the
work, using WWW:Mechanize or some other method for full automation of
these tests might be useful, but I think the generated HTML file is a
step forward from the cut-and-paste solution currently in use.