Beginnings of OptionFields tester.
I've noticed that the PII_Options field doesn't seem to properly contain
the values it should.  This is the beginning of a test file for what
PII_Options should contain.

The initial step here is to simply create one of my forms to create a
test transaction, and verify that the transaction has been created.