Use random string in test transaction name.
This makes the test more easily reproducible and rerunable with the same
sandbox account.  If the options-payment.html file exists, a random
string already embedded in the file is used for the test.

If it doesn't exist, a new random string is generated and the
options-payment.html file is created using it.

This is valuable for when changing this test or wanting rerun all tests
from scratch.