Document and test 2-options-field only PayPal rule
PayPal transactions, may, as it turns out, have only two options fields
each with a key/value pair.  This is not documented well and the fact
that additional options fields "seem to work" at first is a trap for the
unwary.

The API therefore should carefully document this fact, and the tests
should test for this fact explicitly.

The upside is that if PayPal ever supports additional key/value pairs,
the tests will fail and we'll find out that way! :)